pyproject.toml
[project]
name = "robloxscripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keyboard>=0.13.5",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "pyautogui>=0.9.54",
    "pynput>=1.8.1",
]
